rename quality mark to codebase, clean up codebase tests
