+++ setup.cfg
@@ -23,7 +23,7 @@ markers =
     examples: an examples image-diff test
     integration: an integration test
     js: a javascript test
-    quality: a code quality test
+    codebase: a test for codebase quality or policy
     sampledata: a test for bokeh.sampledata
     selenium: a test as requiring selenium
     unit: a python unit test (implicitly assigned for tests otherwise unmarked)
