+++ tests/codebase/__init__.py
@@ -1,0 +1,11 @@
+#-----------------------------------------------------------------------------
+# Copyright (c) 2012 - 2017, Anaconda, Inc. All rights reserved.
+#
+# Powered by the Bokeh Development Team.
+#
+# The full license is in the file LICENSE.txt, distributed with this software.
+#-----------------------------------------------------------------------------
+
+from os.path import abspath, join, pardir, split
+
+TOP_PATH = abspath(join(split(__file__)[0], pardir, pardir))
